- ScratchPads are alive!!
For those addicted to WM's. Having scratchpads is an important feature.
If you want to see it in operation please visit https://youtu.be/t_Gu5cazWzw
If you already live in my dotfiles, you need to install tdrop-git from aur and update sxhkdrc file, check the new code, and update bspwmrc file lines 96 and 97 must be:

bspc rule -a scratch sticky=on state=floating focus=on

If you don't know what it is, well basically it's a window in this case a terminal but it can be any application you want, which will remain active and hidden so you can review it whenever you want.

It is commonly used as a notepad. You can show/hide it with "super + alt + o"

If you wanna add yours or modify the current one, check the man doc for tdrop or tdrop --help

- I have tried to make the configuration more modular and easy to access and easy to edit the themes.
For that reason the configuration files of alacritty, dunst and picom were removed from each theme and now just be one file and the configurations of each theme will be in the theme.sh file.

- Modify the gaps of all the themes I reduced them and adjust the space between the windows and the bars to make them visually perfect and the shadows of the windows do not invade the space of the bars.

- The HideBar script has been updated due to the modifications of the previous point.

- The script to check and list available updates now also shows AUR packages if there are available to update.

- It seemed to me that in Aline's bar the colors of the deskstops were lost, so I modified them a bit, raising the intensity a bit.

- The same for the Melissa theme, I've always been lost and I don't know what desktop I'm on. now they can be identified without problems.

- Modified wallpaper in Aline, Pamela and z0mbi3 rices.

- Updated scientifica font files.

- tdrop (scratchpad) added to paru dependencies. xdotool and xorg-xwininfo added to install dependencies.

- Wallpapers added to Andrea, Melissa and Karla themes.