Improved RiceSelector & Picom settings
Improved default shadows and fading in picom settings.

Yesterday I forgot to make the changes to the RiceSelector. Trying to make my dotfiles more modular, now you just need to add or remove a directory from the ~/.config/bspwm/rices/ directory to add or remove themes respectively.

Obviously, if you are going to add a new theme, copy the structure of another theme as an example and make that your starting point.