Dunst: Update
Dunst has been updated to a new version and they finally introduced dunstctl reload.
It has never been necessary to start Dunst explicitly. However, to load the configuration file it was necessary to kill the process if it was active and restart it.
Now thanks to "dunstctl reload", you can avoid searching if it is active, killing the process and restarting it between each theme change.