All scripts modified.
I've modified all the scripts to make them POSIX compliant. Why? To make them portable, and because in Arch, /bin/sh is linked to Bash.

ls -l /usr/bin/sh

And according to the Arch Wiki, Dash is 35% to 40% faster than Bash, and I've made the change to link sh to dash.

But it's not something that will affect you; the functionality will remain the same. In fact, you may or may not update the scripts. I'm only uploading these changes so I don't have different code on my machine and in the repository, because I'm sure I'll forget in the future and mix up code, and to prevent that from happening.